refactor: consolidate callbacks into callbacks.py and update shell confirmation handling
