+++ src/coding_assistant/agents/tests/test_callbacks_integration.py
@@ -5,7 +5,7 @@ import pytest
 
 from coding_assistant.agents.execution import run_agent_loop
 from coding_assistant.agents.callbacks import NullToolCallbacks
-from coding_assistant.rich_callbacks import ConfirmationToolCallbacks
+from coding_assistant.callbacks import ConfirmationToolCallbacks
 from coding_assistant.agents.tests.helpers import (
     FakeCompleter,
     FakeFunction,
