+++ src/coding_assistant/agents/tests/test_execution.py
@@ -1,7 +1,7 @@
 import pytest
 
 from coding_assistant.agents.callbacks import NullProgressCallbacks, NullToolCallbacks
-from coding_assistant.rich_callbacks import ConfirmationToolCallbacks
+from coding_assistant.callbacks import ConfirmationToolCallbacks
 from coding_assistant.agents.execution import do_single_step, handle_tool_call, handle_tool_calls
 from coding_assistant.agents.tests.helpers import FakeFunction, FakeToolCall, make_test_agent, make_test_context, make_ui_mock
 from coding_assistant.agents.types import AgentDescription, AgentState, AgentContext, TextResult, Tool, ToolResult
