+++ src/coding_assistant/agents/tests/test_execution_shorten_conversation.py
@@ -3,7 +3,7 @@ import json
 import pytest
 
 from coding_assistant.agents.callbacks import NullProgressCallbacks, NullToolCallbacks
-from coding_assistant.rich_callbacks import ConfirmationToolCallbacks
+from coding_assistant.callbacks import ConfirmationToolCallbacks
 from coding_assistant.agents.execution import do_single_step, handle_tool_call
 from coding_assistant.agents.tests.helpers import (
     FakeFunction,
